requirements.txt
# ================================================================
# GOLD TIER - COMPLETE DEPENDENCIES
# Autonomous Employee System
# ================================================================
# Install: pip install -r requirements.txt
# Then:    python -m playwright install chromium
# ================================================================

# ================================================================
# BROWSER AUTOMATION (Facebook, Instagram, Twitter)
# ================================================================
playwright>=1.40.0

# ================================================================
# FILE WATCHING & AUTOMATION
# ================================================================
watchdog>=2.1.0

# ================================================================
# WEB & API REQUESTS
# ================================================================
requests>=2.31.0
urllib3>=2.0.0

# ================================================================
# ENVIRONMENT & CONFIGURATION
# ================================================================
python-dotenv>=1.0.0

# ================================================================
# MCP SERVERS - Flask based
# ================================================================
flask>=2.0.0
flask-cors>=3.0.0

# ================================================================
# JSON & DATA PROCESSING
# ================================================================
jsonschema>=4.0.0

# ================================================================
# LOGGING & AUDIT
# ================================================================
python-json-logger>=2.0.0

# ================================================================
# SCHEDULING & TASK MANAGEMENT
# ================================================================
schedule>=1.2.0
apscheduler>=3.10.0

# ================================================================
# ODOO INTEGRATION (JSON-RPC)
# ================================================================
# (uses requests library, no additional deps)

# ================================================================
# EMAIL (SMTP/IMAP)
# ================================================================
# (built-in Python libraries: smtplib, imaplib)

# ================================================================
# WHATSAPP INTEGRATION
# ================================================================
# (uses requests library for WhatsApp Business API)

# ================================================================
# AI INTEGRATION (Qwen/Dashscope)
# ================================================================
# dashscope>=1.0.0  # Optional - for Qwen AI

# ================================================================
# UTILITY LIBRARIES
# ================================================================
typing-extensions>=4.0.0

# ================================================================
# END OF REQUIREMENTS
# ================================================================
